Database with Cloud Deployement/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
google-cloud-firestore>=2.16.0
